requirements.txt
Django==1.5.2
South==0.8.2
Werkzeug==0.9.3
argparse==1.2.1
django-debug-toolbar==0.9.4
django-extensions==1.1.1
django-reversion==1.7.1
ipdb==0.7
ipython==1.0.0
jedi==0.6.0
psycopg2==2.5.1
six==1.3.0
wsgiref==0.1.2
django-admin-bootstrapped==0.4.3
django-simple-captcha==0.4.0
